refactor: remove components folders
