+++ src/renderer/windows/MainWindow/InstanceContent/ConnectionSelectorContainer/Config/index.scss
