+++ src/renderer/windows/MainWindow/InstanceContent/ConnectionSelectorContainer/Favorite.jsx
