+++ src/renderer/windows/MainWindow/InstanceContent/DatabaseContainer/AddButton/index.scss
