+++ src/renderer/windows/MainWindow/InstanceContent/DatabaseContainer/Content/Config/index.scss
