+++ src/renderer/windows/MainWindow/InstanceContent/DatabaseContainer/Content/Footer.jsx
