+++ src/renderer/windows/MainWindow/InstanceContent/DatabaseContainer/Content/KeyContent/BaseContent/HashContent.jsx
@@ -1,12 +1,12 @@
 'use strict'
 
 import React from 'react'
-import BaseContent from './BaseContent'
+import BaseContent from '.'
 import SplitPane from 'react-split-pane'
 import {Table, Column} from 'fixed-data-table-contextmenu'
 import Editor from './Editor'
-import AddButton from '../../../../AddButton'
-import ContentEditable from '../../../../ContentEditable'
+import AddButton from '../../../AddButton'
+import ContentEditable from '../../../ContentEditable'
 import ReactDOM from 'react-dom'
 import {clipboard, remote} from 'electron'
