+++ src/renderer/windows/MainWindow/InstanceContent/DatabaseContainer/Content/KeyContent/BaseContent/SetContent.jsx
@@ -1,14 +1,14 @@
 'use strict'
 
 import React from 'react'
-import BaseContent from './BaseContent'
+import BaseContent from '.'
 import SplitPane from 'react-split-pane'
 import {Table, Column} from 'fixed-data-table-contextmenu'
 import Editor from './Editor'
-import AddButton from '../../../../AddButton'
+import AddButton from '../../../AddButton'
 import {remote} from 'electron'
 
-require('./BaseContent/index.scss')
+require('./index.scss')
 
 class SetContent extends BaseContent {
   save(value, callback) {
