+++ src/renderer/windows/MainWindow/InstanceContent/DatabaseContainer/Content/KeyContent/BaseContent/ZSetContent.jsx
@@ -1,17 +1,17 @@
 'use strict'
 
 import React from 'react'
-import BaseContent from './BaseContent'
+import BaseContent from '.'
 import SplitPane from 'react-split-pane'
 import {Table, Column} from 'fixed-data-table-contextmenu'
 import Editor from './Editor'
 import SortHeaderCell from './SortHeaderCell'
-import AddButton from '../../../../AddButton'
-import ContentEditable from '../../../../ContentEditable'
+import AddButton from '../../../AddButton'
+import ContentEditable from '../../../ContentEditable'
 import ReactDOM from 'react-dom'
 import {clipboard} from 'electron'
 
-require('./BaseContent/index.scss')
+require('./index.scss')
 
 class ZSetContent extends BaseContent {
   save(value, callback) {
