+++ src/renderer/windows/MainWindow/InstanceContent/DatabaseContainer/Content/TabBar/index.scss
