+++ src/renderer/windows/MainWindow/InstanceContent/DatabaseContainer/Content/Terminal/index.jsx
@@ -5,7 +5,7 @@ import commands from 'redis-commands'
 import splitargs from 'redis-splitargs'
 import 'jquery.terminal'
 
-require('../../../../../../../../../../../../node_modules/jquery.terminal/css/jquery.terminal.css')
+require('../../../../../../../../node_modules/jquery.terminal/css/jquery.terminal.css')
 require('./index.scss')
 
 class Terminal extends React.PureComponent {
