+++ src/renderer/windows/MainWindow/InstanceContent/DatabaseContainer/Content/Terminal/index.scss
