+++ src/renderer/windows/MainWindow/InstanceContent/DatabaseContainer/ContentEditable/index.jsx
