+++ src/renderer/windows/MainWindow/InstanceContent/DatabaseContainer/index.scss
