+++ src/renderer/windows/MainWindow/InstanceContent/Modal/icon.png
