+++ src/renderer/windows/MainWindow/InstanceContent/Modal/index.scss
