+++ src/renderer/windows/MainWindow/InstanceContent/Modal/warning.png
