+++ src/renderer/windows/MainWindow/InstanceContent/index.jsx
@@ -1,9 +1,9 @@
 'use strict'
 
 import React, {PureComponent} from 'react'
-import ConnectionSelectorContainer from './components/ConnectionSelectorContainer'
-import DatabaseContainer from './components/DatabaseContainer'
-import Modal from '../../components/InstanceContent/components/Modal'
+import ConnectionSelectorContainer from './ConnectionSelectorContainer'
+import DatabaseContainer from './DatabaseContainer'
+import Modal from './Modal'
 import ReactCSSTransitionGroup from 'react-addons-css-transition-group'
 
 class InstanceContent extends PureComponent {
