+++ src/renderer/windows/MainWindow/InstanceTabs/Tabs.tsx
