+++ src/renderer/windows/MainWindow/InstanceTabs/index.tsx
