+++ src/renderer/windows/MainWindow/InstanceTabs/main.scss
