+++ src/renderer/windows/MainWindow/index.jsx
@@ -3,8 +3,8 @@
 import React, {PureComponent} from 'react'
 import {createSelector} from 'reselect'
 import {Provider, connect} from 'react-redux'
-import InstanceTabs from './components/InstanceTabs'
-import InstanceContent from './components/InstanceContent'
+import InstanceTabs from './InstanceTabs'
+import InstanceContent from './InstanceContent'
 import DocumentTitle from 'react-document-title'
 import {createInstance, selectInstance, delInstance, moveInstance} from 'Redux/actions'
 import store from 'Redux/store'
